requirements.txt
asgiref==3.7.2
certifi==2023.5.7
charset-normalizer==3.1.0
distro==1.8.0
Django==4.2.2
django-cors-headers==4.1.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
idna==3.4
numpy==1.24.3
openpyxl==3.1.2
pandas==2.0.2
Pillow==9.5.0
pip==22.3.1
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
setuptools==65.5.0
six==1.16.0
sqlparse==0.4.4
tabula-py==2.7.0
tzdata==2023.3
urllib3==2.0.3
mysqlclient